+++ core-ui/src/main/java/com/anytypeio/anytype/core_ui/common/Focusable.kt
@@ -4,5 +4,5 @@ package com.anytypeio.anytype.core_ui.common
  * Defines a view that can be focused.
  */
 interface Focusable {
-    val isFocused: Boolean
+    var isFocused: Boolean
 }